services/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coachella-api"
version = "0.1.0"
description = "API for Coachella Group Planner"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "pydantic-settings>=2.10.0"
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "httpx>=0.28.0",
  "ruff>=0.12.0",
  "mypy>=1.16.0"
]

[tool.pytest.ini_options]
pythonpath = ["app"]

[tool.ruff]
line-length = 100
